fix: Moving componentEnums content to domTypes for better type management
